fix(from): correct next tick method call
Because promises are processed on a microtask queue and timeouts are
processed on the task queue, the promise was resolving _before_ the
`setImmediate` was called. This caused an unhandled rejection the
promise was rejected asyncronously.

By calling the method on the next tick of the microqueue, we avoid this
issue and still retain setting up the stream handlers after the call.